+++ cmd/main.go
@@ -231,6 +231,13 @@ func main() {
 		setupLog.Error(err, "unable to create controller", "controller", "TemporalSchedule")
 		os.Exit(1)
 	}
+	if err := (&controller.TemporalDevServerReconciler{
+		Client: mgr.GetClient(),
+		Scheme: mgr.GetScheme(),
+	}).SetupWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create controller", "controller", "TemporalDevServer")
+		os.Exit(1)
+	}
 	webhooksEnabled := os.Getenv("ENABLE_WEBHOOKS") != "false"
 	if webhooksEnabled {
 		if err := webhookv1alpha1.SetupTemporalClusterWebhookWithManager(mgr); err != nil {